+++ packages/sources/view-function-multi-chain/test/integration/adapter.test.ts
@@ -45,6 +45,10 @@ describe('execute', () => {
     spy.mockRestore()
   })
 
+  afterEach(() => {
+    nock.cleanAll()
+  })
+
   describe('function endpoint', () => {
     it('should return success', async () => {
       const data = {
@@ -63,11 +67,12 @@ describe('execute', () => {
         contract: '0x2c1d072e956AFFC0D435Cb7AC38EF18d24d9127c',
         function: 'function latestAnswer() external view returns (int256)',
         network: 'ethereum_mainnet',
-        data: {
-          decimals: {
+        additionalRequests: [
+          {
+            name: 'decimals',
             signature: 'function decimals() view returns (uint8)',
           },
-        },
+        ],
       }
       mockETHMainnetContractCallResponseSuccess()
       const response = await testAdapter.request(data)
@@ -75,37 +80,21 @@ describe('execute', () => {
       expect(response.json()).toMatchSnapshot()
     })
 
-    it('should return success with additional data requests ', async () => {
+    it('should fail additional data requests in case of missing signature', async () => {
       const data = {
         contract: '0x2c1d072e956AFFC0D435Cb7AC38EF18d24d9127c',
         function: 'function latestAnswer() external view returns (int256)',
         network: 'ethereum_mainnet',
-        data: {
-          decimals: {
-            signature: 'function decimals() view returns (uint8)',
-          },
-        },
-      }
-      mockETHMainnetContractCallResponseSuccess()
-      const response = await testAdapter.request(data)
-      expect(response.statusCode).toBe(200)
-      expect(response.json()).toMatchSnapshot()
-    })
-
-    it('should skip additional data requests in case of missing signature', async () => {
-      const data = {
-        contract: '0x2c1d072e956AFFC0D435Cb7AC38EF18d24d9127c',
-        function: 'function latestAnswer() external view returns (int256)',
-        network: 'ethereum_mainnet',
-        data: {
-          decimals: {
+        additionalRequests: [
+          {
+            name: 'decimals',
             signature: '',
           },
-        },
+        ],
       }
       mockETHMainnetContractCallResponseSuccess()
       const response = await testAdapter.request(data)
-      expect(response.statusCode).toBe(200)
+      expect(response.statusCode).toBe(502)
       expect(response.json()).toMatchSnapshot()
     })